requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
PyMySQL==1.1.0
openai>=1.12.0
reportlab==4.0.4
httpx>=0.24.0,<1.0.0
